# docs/openapi/openapi.json
{
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "responses": {
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ],
        "summary": "Generate access and refresh tokens"
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "responses": {
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Authentication"
        ],
        "summary": "Revoke current token",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/students/": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "sort_by",
            "schema": {
              "type": "string",
              "default": "created_at"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page",
            "description": "Page number (1-indexed)",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "required": false
          },
          {
            "in": "query",
            "name": "per_page",
            "description": "Results per page, max 100",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "required": false
          },
          {
            "in": "query",
            "name": "department",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "division",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "search",
            "description": "Search by name or roll number",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "min_attendance",
            "description": "Filter by min attendance %",
            "schema": {
              "type": "number"
            },
            "required": false
          }
        ],
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "List students",
        "description": "Returns a paginated list of students. Faculty can only see their own division. Admin sees all.",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "responses": {
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "409": {
            "description": "Roll number already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCreate"
              }
            }
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "Create student",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/students/{student_id}": {
      "get": {
        "responses": {
          "404": {
            "description": "Student not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "Get student by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "student_id",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 0
          }
        }
      ],
      "patch": {
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentUpdate"
              }
            }
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "Partially update student",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "Soft-delete student",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/attendance/": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "Page number (1-indexed)",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "required": false
          },
          {
            "in": "query",
            "name": "per_page",
            "description": "Results per page, max 100",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "required": false
          },
          {
            "in": "query",
            "name": "student_id",
            "schema": {
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "subject",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "start_date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "end_date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": false
          }
        ],
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Attendance"
        ],
        "summary": "Get list of attendance records",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "DY Patil University ERP API",
    "version": "v1"
  },
...
  "openapi": "3.1.0",
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Error code"
          },
          "status": {
            "type": "string",
            "description": "Error name"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "object",
            "description": "Errors"
          }
        }
      },
...
      "Login": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "faculty",
              "student"
            ]
          }
        },
        "required": [
          "password",
          "role",
          "username"
        ]
      },
      "Error1": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Not found"
          },
          "code": {
            "type": "string",
            "example": "STUDENT_NOT_FOUND"
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "default": null
          }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "department": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfile"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "user"
        ]
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "per_page": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          },
          "has_next": {
            "type": "boolean"
          },
          "has_prev": {
            "type": "boolean"
          }
        }
      },
      "StudentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "roll": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "division": {
            "type": "string"
          },
          "semester": {
            "type": "integer"
          },
          "phone": {
            "type": "string"
          },
          "attendance_pct": {
            "type": "number",
            "default": 0.0,
            "description": "Current month attendance %"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StudentCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "roll": {
            "type": "string",
            "example": "CS2024001"
          },
          "department": {
            "type": "string",
            "enum": [
              "CS",
              "IT",
              "ENTC",
              "MECH",
              "CIVIL",
              "AIDS",
              "AIML"
            ]
          },
          "year": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4
          },
          "division": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "C",
              "D"
            ]
          },
          "semester": {
            "type": "integer",
            "minimum": 1,
            "maximum": 8
          },
          "dob": {
            "type": "string",
            "format": "date"
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+?[0-9]{10,13}$"
          },
          "parent_phone": {
            "type": "string",
            "pattern": "^\\+?[0-9]{10,13}$"
          },
          "address": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "department",
          "division",
          "dob",
          "email",
          "name",
          "roll",
          "semester",
          "year"
        ]
      },
      "StudentUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
      "DEFAULT_ERROR": {
        "description": "Default error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UNPROCESSABLE_ENTITY": {
        "description": "Unprocessable Entity",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}